MVVM: Create `FeedRefreshViewController` to separate refresh feed responsibilities
