+++ MVVM Module/Source/Feed UI/Controllers/FeedViewController.swift
@@ -5,11 +5,7 @@
 import UIKit
 
 public final class FeedViewController: UITableViewController, UITableViewDataSourcePrefetching {
-	@IBOutlet weak var errorView: ErrorView!
-
-	var viewModel: FeedViewModel? {
-		didSet { bind() }
-	}
+	@IBOutlet var refreshController: FeedRefreshViewController?
 	
 	var tableModel = [FeedImageCellController]() {
 		didSet { tableView.reloadData() }
@@ -17,31 +13,8 @@ public final class FeedViewController: UITableViewController, UITableViewDataSou
 	
 	public override func viewDidLoad() {
 		super.viewDidLoad()
-		
-		refresh()
-	}
-	
-	@IBAction private func refresh() {
-		viewModel?.loadFeed()
-	}
-	
-	func bind() {
-		title = viewModel?.title
-		viewModel?.onLoadingStateChange = { [weak self] isLoading in
-			if isLoading {
-				self?.refreshControl?.beginRefreshing()
-			} else {
-				self?.refreshControl?.endRefreshing()
-			}
-		}
 
-		viewModel?.onErrorStateChange = { [weak self] errorMessage in
-			if let message = errorMessage {
-				self?.errorView.show(message: message)
-			} else {
-				self?.errorView.hideMessage()
-			}
-		}
+		refreshController?.refresh()
 	}
 	
 	public override func viewDidLayoutSubviews() {
